Fixed several warnings in the 64-bit MSVC 2015 build.
Some of these probably also affect other MSVC bulids (32-bit, 2010):
- Fixed int casting for strlen() and strnlen() uses.
- TextFuncs string conversion functions:
  - Use -1 instead of strlen() or wcslen() when using passing a
    NULL-terminated string.
  - Cast string::size() to int when converting a C++ string.
- KeyManager::processConfigLine(): Use unsigned int instead of size_t.
- CisoGcnReader, WbfsReader:
  - Use unsigned int instead of uint16_t to store temporary logical and
    physical block numbers. This reduces the number of casts, and it may
    slightly improve performance on x86, since x86 handles 32-bit
    operations faster than 16-bit.
  - Marked temporary block number variables as const.
  - WbfsReader::ALIGN_LBA(): Cast the inversion to size_t.
  - WbfsReader: Added some uint16_t casts where necessary.
- AesCipherTest: AesCipher::decrypt() and CompareByteArrays() both take
  unsigned int, not size_t.
- RP_ShellPropSheetExt: Cast control identifiers to (HMENU)(INT_PTR)
  instead of just (HMENU) to work around a warning about casting a 32-bit
  integer to a 64-bit pointer:
  - RP_ShellPropSheetExt.cpp(411): warning C4312: 'type cast': conversion from 'int' to 'HMENU' of greater size

+++ src/libromdata/DMG.cpp
@@ -565,9 +565,8 @@ int DMG::loadFieldData(void)
 	 * 
 	 * Current method is the first one.
 	 */
-	len = strnlen(romHeader->title,(romHeader->cgbflag < 0x80 ? 16 : 15));
-	
-	
+	len = (int)strnlen(romHeader->title,(romHeader->cgbflag < 0x80 ? 16 : 15));
+
 	if ((romHeader->cgbflag&0x3F) == 0 &&		// gameid is only present when cgbflag is
 	    4 == strnlen(romHeader->title+11,4)){	// gameid is always 4 characters long
 		m_fields->addData_string(latin1_to_rp_string(romHeader->title,std::min(len,11)));
